+++ extensions/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support for the upcoming repository URL feature for `PhylumApi.create_project`
+- `generateLockfiles` parameter for `parseLockfile` to inhibit lockfile generation
 
 ## 5.7.2 - 2023-10-10