+++ extensions/CHANGELOG.md
@@ -8,12 +8,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- `skipSandbox` parameter for `parseLockfile` to generate lockfiles without sandbox protection
+- `generateLockfiles` parameter for `parseLockfile` to inhibit lockfile generation
+
 ## 5.8.0 - 2023-10-24
 
 ### Added
 
 - Support for the upcoming repository URL feature for `PhylumApi.create_project`
-- `generateLockfiles` parameter for `parseLockfile` to inhibit lockfile generation
 
 ## 5.7.2 - 2023-10-10